// src/types/bancos.ts
export type Bancos = {
    bank_name: string;
    acronym_bank: string;
    cards: string;
    status: string;
    dados: {
        id: string;
        bank_name: string;
        bank_code: string;
        bank_color: string;
        acronym_bank: string;
        created_at: string;
        status: string
    }
}